requirements.txt
PyQt5==5.15.9
opencv-python==4.8.1.78
ultralytics==8.0.196
easyocr==1.7.0
psycopg2-binary==2.9.9
numpy==1.24.3
Pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
pyinstaller==6.1.0